camelCase instead of PascalCase for package name
