+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/mainMethodExamples/invalid/Name.java
@@ -2,11 +2,11 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package de.uni_passau.fim.se2.pipeline_helper.checkers.MainMethodExamples.valid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.invalid;
 
 @SuppressWarnings("checkstyle:HideUtilityClassConstructor")
-public class PublicStaticStringArrayParam {
+public class Name {
 
-    public static void main(String[] args) {
+    public static void myMain(String[] args) {
     }
 }
