+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/mainMethodExamples/package-info.java
@@ -5,4 +5,4 @@
 /**
  * Java Classes with different "main" methods, which are valid or invalid for the checker.
  */
-package de.uni_passau.fim.se2.pipeline_helper.checkers.MainMethodExamples;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples;
