+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/mainMethodExamples/valid/PackagePrivateStaticStringArrayParam.java
@@ -2,11 +2,11 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package de.uni_passau.fim.se2.pipeline_helper.checkers.MainMethodExamples.invalid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.valid;
 
 @SuppressWarnings("checkstyle:HideUtilityClassConstructor")
-public class Name {
+public class PackagePrivateStaticStringArrayParam {
 
-    public static void myMain(String[] args) {
+    static void main(String[] args) {
     }
 }
