+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/mainMethodExamples/valid/ProtectedInstanceStringArrayParam.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package de.uni_passau.fim.se2.pipeline_helper.checkers.MainMethodExamples.valid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.valid;
 
 public class ProtectedInstanceStringArrayParam {
