+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/mainMethodExamples/valid/ProtectedStaticNoParam.java
@@ -2,12 +2,11 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package de.uni_passau.fim.se2.pipeline_helper.checkers.MainMethodExamples.invalid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.valid;
 
 @SuppressWarnings("checkstyle:HideUtilityClassConstructor")
-public class ParameterCount {
+public class ProtectedStaticNoParam {
 
-    static void main(String[] args, int count) {
-
+    protected static void main() {
     }
 }
