+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/mainMethodExamples/valid/package-info.java
@@ -5,4 +5,4 @@
 /**
  * Classes containing a valid main method for the checker.
  */
-package de.uni_passau.fim.se2.pipeline_helper.checkers.MainMethodExamples.valid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.valid;
